pyproject.toml
[tool.poetry]
name = "graphql"
version = "0.1.0"
description = ""
authors = ["Prabal Pathak <prabal01pathak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
azure-messaging-webpubsubservice = "^1.0.1"
websockets = "^10.4"
python-dotenv = "^0.21.1"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sse-starlette = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
